Reorganize sass file structure;
