+++ src/actions/index.js
@@ -1,11 +1,11 @@
 export const showTracks = (tracks) => ({
-    type: 'SHOW_TRACKS',
-    tracks
+  type: 'SHOW_TRACKS',
+  tracks
 })
 
 export const isLoading = (bool) => {
-	return {
-		type: 'IS_LOADING',
-		isLoading: bool
-	}
+  return {
+    type: 'IS_LOADING',
+    isLoading: bool
+  }
 }
