+++ src/components/App.js
@@ -1,15 +1,13 @@
 import React from 'react';
-import './App.scss';
 import { connect } from 'react-redux'
 import Header from '../containers/Header';
-import Main from '../containers/Main';
+import Main from '../containers/TrackContainer';
 
 
 function App() {
   return (
     <div className="App">
       <Header />
-      {/* <Search /> */}
       <Main />
     </div>
   );
