--- src/components/App.scss
@@ -1,9 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Quicksand&display=swap');
-
-.App {
-  text-align: center;
-  font-family: 'Quicksand', sans-serif;
-  background-color: #CEE0DC;
-  margin-top: -24px;
-  padding: 0px;
-}
