--- src/containers/Header.scss
@@ -1,3 +1,0 @@
-h1 {
-  padding: 20px;
-}
