+++ src/index.js
@@ -8,14 +8,15 @@ import { createStore, applyMiddleware } from 'redux'
 import rootReducer from './reducers'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import thunk from 'redux-thunk';
+import '../src/styles/main.scss'
 
 
 const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)))
 
 const router = (
-	<Provider store={store}>
-	  <App />
-	</Provider>
+  <Provider store={store}>
+    <App />
+  </Provider>
 )
 
 ReactDOM.render(router, document.getElementById('root'));
