+++ src/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import {tracksReducer, isLoading} from './tracksReducer';
+import { tracksReducer, isLoading } from './tracksReducer';
 
 
 const rootReducer = combineReducers({
