+++ src/styles/_header.scss
@@ -1,0 +1,4 @@
+h1 {
+  padding: 20px;
+  font-family: $font;
+}
