+++ src/styles/_variables.scss
@@ -1,0 +1,4 @@
+@import url("https://fonts.googleapis.com/css?family=Quicksand&display=swap");
+
+$font: "Quicksand", sans-serif;
+$light-blue: #cee0dc;
