+++ src/styles/main.scss
@@ -1,0 +1,4 @@
+@import "/variables";
+
+@import "/header";
+@import "/track";
